requirements.txt
streamlit==1.28.1
openai==1.12.0
python-dotenv==0.21.0
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
langchain==0.1.0
httpx==0.24.1
